test: migrate ignore.rs unit tests to integration tests
