+++ src/authorship/ignore.rs
@@ -288,545 +288,3 @@ fn split_gitattributes_tokens(line: &str) -> Vec<String> {
 
     tokens
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::git::repository::from_bare_repository;
-    use crate::git::test_utils::TmpRepo;
-    use std::fs;
-    use std::path::Path;
-    use std::process::Command;
-
-    fn run_git(cwd: &Path, args: &[&str]) {
-        crate::git::test_utils::init_test_git_config();
-        let output = Command::new(crate::config::Config::get().git_cmd())
-            .args(args)
-            .current_dir(cwd)
-            .output()
-            .expect("git command should run");
-        assert!(
-            output.status.success(),
-            "git {:?} failed:\nstdout: {}\nstderr: {}",
-            args,
-            String::from_utf8_lossy(&output.stdout),
-            String::from_utf8_lossy(&output.stderr)
-        );
-    }
-
-    fn make_bare_repo(
-        root_gitattributes: Option<&str>,
-        parent_gitattributes: Option<&str>,
-    ) -> (tempfile::TempDir, Repository) {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let source = temp.path().join("source");
-        let bare = temp.path().join("bare.git");
-        fs::create_dir_all(&source).expect("create source");
-
-        run_git(&source, &["init"]);
-        run_git(&source, &["config", "user.name", "Test User"]);
-        run_git(&source, &["config", "user.email", "test@example.com"]);
-
-        fs::write(source.join("README.md"), "# repo\n").expect("write readme");
-        if let Some(attrs) = root_gitattributes {
-            fs::write(source.join(".gitattributes"), attrs).expect("write attrs");
-        }
-
-        run_git(&source, &["add", "."]);
-        run_git(&source, &["commit", "-m", "initial"]);
-        run_git(
-            temp.path(),
-            &[
-                "clone",
-                "--bare",
-                source.to_str().unwrap(),
-                bare.to_str().unwrap(),
-            ],
-        );
-
-        if let Some(parent_attrs) = parent_gitattributes {
-            fs::write(temp.path().join(".gitattributes"), parent_attrs)
-                .expect("write parent attrs");
-        }
-
-        (
-            temp,
-            from_bare_repository(&bare).expect("bare repository should load"),
-        )
-    }
-
-    #[test]
-    fn defaults_include_snapshot_and_lock_patterns() {
-        let defaults = default_ignore_patterns();
-        assert!(defaults.contains(&"**/*.snap".to_string()));
-        assert!(defaults.contains(&"Cargo.lock".to_string()));
-        assert!(defaults.contains(&"*.generated.*".to_string()));
-    }
-
-    #[test]
-    fn defaults_ignore_drizzle_meta_files() {
-        let defaults = default_ignore_patterns();
-        let matcher = build_ignore_matcher(&defaults);
-
-        assert!(should_ignore_file_with_matcher(
-            "web/drizzle/meta/_journal.json",
-            &matcher
-        ));
-        assert!(should_ignore_file_with_matcher(
-            "web/drizzle/meta/0001_snapshot.json",
-            &matcher
-        ));
-        assert!(should_ignore_file_with_matcher(
-            "drizzle/meta/0032_snapshot.json",
-            &matcher
-        ));
-        // Should not ignore non-meta drizzle files
-        assert!(!should_ignore_file_with_matcher(
-            "drizzle/0001_initial.sql",
-            &matcher
-        ));
-    }
-
-    #[test]
-    fn defaults_do_not_ignore_generic_snapshots_directories() {
-        let defaults = default_ignore_patterns();
-        let matcher = build_ignore_matcher(&defaults);
-
-        assert!(!should_ignore_file_with_matcher(
-            "backups/snapshots/state.json",
-            &matcher
-        ));
-        assert!(should_ignore_file_with_matcher(
-            "tests/__snapshots__/feature.snap",
-            &matcher
-        ));
-        assert!(should_ignore_file_with_matcher(
-            "tests/snapshots/feature.snap",
-            &matcher
-        ));
-    }
-
-    #[test]
-    fn defaults_ignore_nested_named_lockfiles() {
-        let defaults = default_ignore_patterns();
-        let matcher = build_ignore_matcher(&defaults);
-
-        assert!(should_ignore_file_with_matcher(
-            "apps/web/Gemfile.lock",
-            &matcher
-        ));
-        assert!(should_ignore_file_with_matcher(
-            "services/api/package-lock.json",
-            &matcher
-        ));
-        assert!(should_ignore_file_with_matcher(
-            "libs/core/Cargo.lock",
-            &matcher
-        ));
-    }
-
-    #[test]
-    fn should_ignore_file_matches_path_and_filename() {
-        let patterns = vec!["*.lock".to_string(), "**/node_modules/**".to_string()];
-        let matcher = build_ignore_matcher(&patterns);
-        assert!(should_ignore_file("Cargo.lock", &patterns));
-        assert!(should_ignore_file("backend/Cargo.lock", &patterns));
-        assert!(should_ignore_file_with_matcher("Cargo.lock", &matcher));
-        assert!(should_ignore_file_with_matcher(
-            "backend/Cargo.lock",
-            &matcher
-        ));
-        assert!(should_ignore_file(
-            "web/node_modules/lodash/index.js",
-            &patterns
-        ));
-        assert!(should_ignore_file_with_matcher(
-            "web/node_modules/lodash/index.js",
-            &matcher
-        ));
-        assert!(!should_ignore_file("src/main.rs", &patterns));
-        assert!(!should_ignore_file_with_matcher("src/main.rs", &matcher));
-    }
-
-    #[test]
-    fn loads_positive_linguist_generated_only() {
-        let tmp_repo = TmpRepo::new().expect("tmp repo");
-        tmp_repo
-            .write_file(
-                ".gitattributes",
-                "\
-*.generated.ts linguist-generated=true
-dist/** linguist-generated
-vendor/** -linguist-generated
-manual/** linguist-generated=false
-flags/** linguist-generated=1
-other/** linguist-generated=0
-generated\\ files/** linguist-generated=true
-",
-                true,
-            )
-            .expect("write attrs");
-        tmp_repo
-            .commit_with_message("add gitattributes")
-            .expect("commit");
-
-        let patterns =
-            load_linguist_generated_patterns_from_root_gitattributes(tmp_repo.gitai_repo());
-        assert!(patterns.contains(&"*.generated.ts".to_string()));
-        assert!(patterns.contains(&"dist/**".to_string()));
-        assert!(patterns.contains(&"flags/**".to_string()));
-        assert!(patterns.contains(&"generated files/**".to_string()));
-        assert!(!patterns.contains(&"vendor/**".to_string()));
-        assert!(!patterns.contains(&"manual/**".to_string()));
-        assert!(!patterns.contains(&"other/**".to_string()));
-    }
-
-    #[test]
-    fn ignores_gitattributes_macro_definitions() {
-        let tmp_repo = TmpRepo::new().expect("tmp repo");
-        tmp_repo
-            .write_file(
-                ".gitattributes",
-                "\
-[attr]generated linguist-generated=true
-generated/** linguist-generated=true
-",
-                true,
-            )
-            .expect("write attrs");
-        tmp_repo
-            .commit_with_message("add gitattributes")
-            .expect("commit");
-
-        let patterns =
-            load_linguist_generated_patterns_from_root_gitattributes(tmp_repo.gitai_repo());
-
-        assert!(patterns.contains(&"generated/**".to_string()));
-        assert!(!patterns.contains(&"[attr]generated".to_string()));
-    }
-
-    #[test]
-    fn invalid_patterns_fallback_to_exact_path_or_filename() {
-        let patterns = vec!["[".to_string(), "docs/[bad".to_string()];
-        let matcher = build_ignore_matcher(&patterns);
-
-        assert!(should_ignore_file_with_matcher("[", &matcher));
-        assert!(should_ignore_file_with_matcher("docs/[bad", &matcher));
-        assert!(!should_ignore_file_with_matcher("docs/good.rs", &matcher));
-    }
-
-    #[test]
-    fn loads_linguist_generated_from_bare_repo_head() {
-        let (_tmp, bare_repo) = make_bare_repo(
-            Some("generated/** linguist-generated=true\nmanual/** linguist-generated=false\n"),
-            None,
-        );
-
-        let patterns = load_linguist_generated_patterns_from_root_gitattributes(&bare_repo);
-        assert!(patterns.contains(&"generated/**".to_string()));
-        assert!(!patterns.contains(&"manual/**".to_string()));
-    }
-
-    #[test]
-    fn bare_repo_does_not_read_parent_directory_gitattributes() {
-        let (_tmp, bare_repo) = make_bare_repo(None, Some("leak/** linguist-generated=true\n"));
-
-        let patterns = load_linguist_generated_patterns_from_root_gitattributes(&bare_repo);
-        assert!(patterns.is_empty());
-    }
-
-    #[test]
-    fn loads_git_ai_ignore_patterns_from_workdir() {
-        let tmp_repo = TmpRepo::new().expect("tmp repo");
-        tmp_repo
-            .write_file(
-                ".git-ai-ignore",
-                "\
-# This is a comment
-docs/**
-*.pdf
-
-assets/images/**
-",
-                true,
-            )
-            .expect("write .git-ai-ignore");
-        tmp_repo
-            .commit_with_message("add .git-ai-ignore")
-            .expect("commit");
-
-        let patterns = load_git_ai_ignore_patterns(tmp_repo.gitai_repo());
-        assert_eq!(patterns.len(), 3);
-        assert!(patterns.contains(&"docs/**".to_string()));
-        assert!(patterns.contains(&"*.pdf".to_string()));
-        assert!(patterns.contains(&"assets/images/**".to_string()));
-    }
-
-    #[test]
-    fn git_ai_ignore_skips_comments_and_blank_lines() {
-        let tmp_repo = TmpRepo::new().expect("tmp repo");
-        tmp_repo
-            .write_file(
-                ".git-ai-ignore",
-                "\
-# comment line
-   # indented comment
-
-  *.log  
-build/**
-",
-                true,
-            )
-            .expect("write .git-ai-ignore");
-        tmp_repo
-            .commit_with_message("add .git-ai-ignore")
-            .expect("commit");
-
-        let patterns = load_git_ai_ignore_patterns(tmp_repo.gitai_repo());
-        assert_eq!(patterns.len(), 2);
-        assert!(patterns.contains(&"*.log".to_string()));
-        assert!(patterns.contains(&"build/**".to_string()));
-    }
-
-    #[test]
-    fn git_ai_ignore_deduplicates_patterns() {
-        let tmp_repo = TmpRepo::new().expect("tmp repo");
-        tmp_repo
-            .write_file(
-                ".git-ai-ignore",
-                "\
-*.pdf
-docs/**
-*.pdf
-",
-                true,
-            )
-            .expect("write .git-ai-ignore");
-        tmp_repo
-            .commit_with_message("add .git-ai-ignore")
-            .expect("commit");
-
-        let patterns = load_git_ai_ignore_patterns(tmp_repo.gitai_repo());
-        assert_eq!(patterns.len(), 2);
-    }
-
-    #[test]
-    fn git_ai_ignore_returns_empty_when_file_missing() {
-        let tmp_repo = TmpRepo::new().expect("tmp repo");
-        tmp_repo
-            .write_file("README.md", "# repo\n", true)
-            .expect("write readme");
-        tmp_repo.commit_with_message("initial").expect("commit");
-
-        let patterns = load_git_ai_ignore_patterns(tmp_repo.gitai_repo());
-        assert!(patterns.is_empty());
-    }
-
-    #[test]
-    fn effective_patterns_include_git_ai_ignore() {
-        let tmp_repo = TmpRepo::new().expect("tmp repo");
-        tmp_repo
-            .write_file(".git-ai-ignore", "custom/**\n*.secret\n", true)
-            .expect("write .git-ai-ignore");
-        tmp_repo
-            .commit_with_message("add .git-ai-ignore")
-            .expect("commit");
-
-        let patterns = effective_ignore_patterns(tmp_repo.gitai_repo(), &[], &[]);
-        assert!(patterns.contains(&"custom/**".to_string()));
-        assert!(patterns.contains(&"*.secret".to_string()));
-        // Default patterns should still be present
-        assert!(patterns.contains(&"*.lock".to_string()));
-    }
-
-    #[test]
-    fn effective_patterns_union_gitattributes_and_git_ai_ignore() {
-        let tmp_repo = TmpRepo::new().expect("tmp repo");
-        tmp_repo
-            .write_file(
-                ".gitattributes",
-                "generated/** linguist-generated=true\n",
-                true,
-            )
-            .expect("write .gitattributes");
-        tmp_repo
-            .write_file(".git-ai-ignore", "docs/**\n", true)
-            .expect("write .git-ai-ignore");
-        tmp_repo
-            .commit_with_message("add gitattributes and git-ai-ignore")
-            .expect("commit");
-
-        let patterns = effective_ignore_patterns(tmp_repo.gitai_repo(), &[], &[]);
-        // From .gitattributes linguist-generated
-        assert!(patterns.contains(&"generated/**".to_string()));
-        // From .git-ai-ignore
-        assert!(patterns.contains(&"docs/**".to_string()));
-        // Defaults
-        assert!(patterns.contains(&"*.lock".to_string()));
-    }
-
-    #[test]
-    fn effective_patterns_union_git_ai_ignore_and_user_patterns() {
-        let tmp_repo = TmpRepo::new().expect("tmp repo");
-        tmp_repo
-            .write_file(".git-ai-ignore", "docs/**\n", true)
-            .expect("write .git-ai-ignore");
-        tmp_repo
-            .commit_with_message("add .git-ai-ignore")
-            .expect("commit");
-
-        let user = vec!["tests/**".to_string()];
-        let patterns = effective_ignore_patterns(tmp_repo.gitai_repo(), &user, &[]);
-        // From .git-ai-ignore
-        assert!(patterns.contains(&"docs/**".to_string()));
-        // From user --ignore flag
-        assert!(patterns.contains(&"tests/**".to_string()));
-        // Defaults
-        assert!(patterns.contains(&"*.lock".to_string()));
-    }
-
-    fn make_bare_repo_with_ignore(
-        root_gitattributes: Option<&str>,
-        git_ai_ignore: Option<&str>,
-    ) -> (tempfile::TempDir, Repository) {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let source = temp.path().join("source");
-        let bare = temp.path().join("bare.git");
-        fs::create_dir_all(&source).expect("create source");
-
-        run_git(&source, &["init"]);
-        run_git(&source, &["config", "user.name", "Test User"]);
-        run_git(&source, &["config", "user.email", "test@example.com"]);
-
-        fs::write(source.join("README.md"), "# repo\n").expect("write readme");
-        if let Some(attrs) = root_gitattributes {
-            fs::write(source.join(".gitattributes"), attrs).expect("write attrs");
-        }
-        if let Some(ignore) = git_ai_ignore {
-            fs::write(source.join(".git-ai-ignore"), ignore).expect("write .git-ai-ignore");
-        }
-
-        run_git(&source, &["add", "."]);
-        run_git(&source, &["commit", "-m", "initial"]);
-        run_git(
-            temp.path(),
-            &[
-                "clone",
-                "--bare",
-                source.to_str().unwrap(),
-                bare.to_str().unwrap(),
-            ],
-        );
-
-        (
-            temp,
-            from_bare_repository(&bare).expect("bare repository should load"),
-        )
-    }
-
-    #[test]
-    fn loads_git_ai_ignore_from_bare_repo_head() {
-        let (_tmp, bare_repo) = make_bare_repo_with_ignore(None, Some("docs/**\n*.pdf\n"));
-
-        let patterns = load_git_ai_ignore_patterns(&bare_repo);
-        assert!(patterns.contains(&"docs/**".to_string()));
-        assert!(patterns.contains(&"*.pdf".to_string()));
-    }
-
-    #[test]
-    fn bare_repo_returns_empty_when_git_ai_ignore_missing() {
-        let (_tmp, bare_repo) = make_bare_repo_with_ignore(None, None);
-
-        let patterns = load_git_ai_ignore_patterns(&bare_repo);
-        assert!(patterns.is_empty());
-    }
-
-    #[test]
-    fn bare_repo_effective_patterns_union_gitattributes_and_git_ai_ignore() {
-        let (_tmp, bare_repo) = make_bare_repo_with_ignore(
-            Some("generated/** linguist-generated=true\n"),
-            Some("docs/**\n"),
-        );
-
-        let patterns = effective_ignore_patterns(&bare_repo, &[], &[]);
-        assert!(patterns.contains(&"generated/**".to_string()));
-        assert!(patterns.contains(&"docs/**".to_string()));
-        assert!(patterns.contains(&"*.lock".to_string()));
-    }
-
-    #[test]
-    fn defaults_include_protobuf_generated_patterns() {
-        let defaults = default_ignore_patterns();
-        // Objective-C protobuf
-        assert!(defaults.contains(&"*.pbobjc.h".to_string()));
-        assert!(defaults.contains(&"*.pbobjc.m".to_string()));
-        // Go protobuf
-        assert!(defaults.contains(&"*.pb.go".to_string()));
-        // C++ protobuf
-        assert!(defaults.contains(&"*.pb.h".to_string()));
-        assert!(defaults.contains(&"*.pb.cc".to_string()));
-        // Python protobuf
-        assert!(defaults.contains(&"*_pb2.py".to_string()));
-        assert!(defaults.contains(&"*_pb2_grpc.py".to_string()));
-        // Swift protobuf
-        assert!(defaults.contains(&"*.pb.swift".to_string()));
-        // Dart protobuf
-        assert!(defaults.contains(&"*.pb.dart".to_string()));
-    }
-
-    #[test]
-    fn defaults_ignore_protobuf_generated_files() {
-        let defaults = default_ignore_patterns();
-        let matcher = build_ignore_matcher(&defaults);
-
-        // Bare filenames
-        assert!(should_ignore_file_with_matcher(
-            "Message.pbobjc.h",
-            &matcher
-        ));
-        assert!(should_ignore_file_with_matcher(
-            "Message.pbobjc.m",
-            &matcher
-        ));
-        assert!(should_ignore_file_with_matcher("service.pb.go", &matcher));
-        assert!(should_ignore_file_with_matcher("message.pb.h", &matcher));
-        assert!(should_ignore_file_with_matcher("message.pb.cc", &matcher));
-        assert!(should_ignore_file_with_matcher("types_pb2.py", &matcher));
-        assert!(should_ignore_file_with_matcher(
-            "service_pb2_grpc.py",
-            &matcher
-        ));
-        assert!(should_ignore_file_with_matcher(
-            "message.pb.swift",
-            &matcher
-        ));
-        assert!(should_ignore_file_with_matcher("message.pb.dart", &matcher));
-
-        // Nested paths
-        assert!(should_ignore_file_with_matcher(
-            "proto/gen/service.pb.go",
-            &matcher
-        ));
-        assert!(should_ignore_file_with_matcher(
-            "ios/Proto/Message.pbobjc.h",
-            &matcher
-        ));
-        assert!(should_ignore_file_with_matcher(
-            "backend/api/types_pb2.py",
-            &matcher
-        ));
-        assert!(should_ignore_file_with_matcher(
-            "cpp/protos/message.pb.cc",
-            &matcher
-        ));
-
-        // Non-protobuf files should NOT be matched
-        assert!(!should_ignore_file_with_matcher("main.go", &matcher));
-        assert!(!should_ignore_file_with_matcher("service.py", &matcher));
-        assert!(!should_ignore_file_with_matcher("header.h", &matcher));
-        assert!(!should_ignore_file_with_matcher("source.cc", &matcher));
-        assert!(!should_ignore_file_with_matcher("app.swift", &matcher));
-        assert!(!should_ignore_file_with_matcher("widget.dart", &matcher));
-        assert!(!should_ignore_file_with_matcher("Objective.m", &matcher));
-    }
-}
